Adds initial category filter
The category filter currently acts as both the category (calendar)
filter and the custom search filter. We may need to split these apart
depending on the needs of the API and UI. I'm going to hold off on any
additional changes to this filter until things are set in stone a bit
more.