Adds initial CRUD UI for new events
Events will be added in a 3-step process. First the event details will
be added, followed by the promotional info, and finally previewing and
publishing.

This commit handles just the event details. This also does not include
many of the UX details like adding multiple events at once or selecting
the price. Those will be added in separate pull requests.