Updates to the API data and models
Most of this came out of a discussion last week about the API. Since
aren't building the API, I'm using a lot of this to make proposals for
what I think it should be.

This also adds an EventInstance model which is used to represent other
dates that events are taking place.