Adds support for uploading images
Images are being uploaded after the event is created so that the initial
POST request does not get hung up by large image uploads. I'm not 100%
sure if this is the desired API behavior, but it seems like a good first
pass to get image uploads working.

The mirage server cannot handle image uploads as form data, so we need
to ignore it. I tested this with a real Rails API to confirm that the
data is being sent correctly to the server.