Adds ability for users to select listservs
When a user is creating a new event, they can now select from a list of
listservs to post their event to.

There is a known bug that hides the list, even if it has items checked
on it. This happens if you select listservs, then change something and
come back to the promotion page. It will works, but the list is
unchecked.