Adds support for selecting and adding new venues
When a user is creating an event, this gives them the ability to search
for existing venues, or add their own. If an existing venue is selected,
the venueId is set and send with the API request. If an existing venue
is not selected, it sends the new venue properties as attributes nested
within a "venue" JSON key.