Adds support for editing events
This required a larger design change because we were sharing the event
model for events and event instances. This became an issue when we
needed to save and edit events because we were using the instance ID
instead of the event ID.

Rather than continue to hack around that, I changed the index and show
event pages to use the event-instance model, and event management is
done through the event model. The instances also use the event-instance
model for their related instances, rather than a raw attribute.