Adds support for previewing events
When a user creates a new event, this lets them preview the card that
will display on the event listing.

The reason I justified the duplication in the markup was because I
wanted to remove the links, and be able to override the values and show
temporary placeholder values in the preview.