Updates event cost selector to use a tabbed interface
We were using radio buttons before, but the designs call for tabs. This
changes it to use the bootstrap tabbed UI. We'll need to style them, but
this gets the functionality to match what the design is asking for.

This also adds a 4th option "Not Applicable". If the cost type is blank,
we'll display that as N/A to users.

Finally, we're pre-selecting "free" as the default type, but this can
change later if needed. There was some back and forth as to what the
most common type would be, so I'm going with this for now.