Adds support for displaying similar content on events
The event show page displays a list of similar content. This content
might be other events or links to other content in the system that are
not events. If it's an event, we link directly to it. If it's not an
event, we link to the external URL provided by the API.