Adds support for front_end_builds
front_end_builds is a Ruby gem that allows us to mount the ember app
within an existing Rails app. With this addon, we can build locally, and
deploy the assets to S3. Once the assets on are on S3, it makes an API
request to the server to notify of the new build available.

    ember deploy --environment=production

Right now this is using my personal AWS account. When we change the
bucket and user, we'll need to update the following:

1. Brocfile.js - finterprint:prepend
2. config/deploy.js - assets:accessKeyId
3. config/deploy.js - assets:secretAccessKey
4. config/deploy.js - assets:bucket