Removes persistence between event creation steps
The API is not setup to support this because once an event is created,
it is automatically published. Rather than changing the API, this
updates things so it only persists when a user publishes the event.

I also fixed an issue where the event UI was busted, and improved the
headers when editing an event.