Fixes date filtering
This fixes an issue where the date was being submitted as "undefined" if
you didn't select a new date. It was also not retaining the dates from
the query params.

I think this was introduced in f6e43cf92a052731c1eeedf61dabbf0293f42ee4
and I must not have QA'd it enough.