Fixes some UX issues with the filters
* getCurrentUser returns promise to ensure the current user is available
  before any other content.
* Fixes date filter so it always shows the correct display text
  depending on which dates have been chosen.
* Use the current user's location for the location filter

One piece of slightly confusing logic is that we are now defaulting to
Today for logged out users and This Week for logged in users. I believe
the reason for this is because logged out users get results for the
entire Upper Valley, but logged in users are associated with a more
narrow region.