Updates map location logic
We had assumed that all events would have latitude and longitude. It
turns out only a handful of them have it, so instead we're using the
physical address of the venue. In most cases, we'll have a "venue locate
name" field, but fall back to the concatenated addresses during the
preview.

The Google Maps version we were using did not seem to support physical
addresses, so I switched it over to embedded maps which use an iframe.
This is how their current app works, and they are ok with continuing to
use this method.