Fixes 'Index out of range' error on staging
For some reason when using the event filter to search for events, it was
randomly raising an 'Index out of range' error and breaking the app. I
wasn't able to track down an issue, but I've fixed it by not relying on
Ember.computed.sort, and instead sorting the array manually.