Switches event time selection to select options
The freeform input was a little confusing, even with the input mask.
This will make it easier to use on mobile, and provides a UI element
that users will be more familiar with.

Ended up commenting out the component spec because it was failing and I
couldn't figure out why. Getting the error:

    Property set failed: object in path "instance"
    could not be found or was destroyed.