Adds validation to event instance end time
We do not want users to submit invalid dates. The API returns an error
in this case, but the error message is null, so we need to implement it
ourselves.

Due to the complexity of having multiple instances, I decided to move
the validation logic into the component, and bubble up the isValid
state. There seems to be some slight performance issues here because of
all the observers and properties around the start and end dates.
Something to address in the future.