Fixes UX for redirecting for creating events
When a user is logged out and clicks the "Create event" button, it was
confusing because we were setting the redirect URL to the current page
rather than the page they were intending to go to.

This fixes it by taking them to the new event page after they log in.