Hides drag hover state of summernote wysiwyg editor
When dragging an image onto the page, the WYSIWYG editor was showing a
"Drag and drop image here" message. We aren't supporting this feature,
but there's no way to disable it without breaking drag & drop
functionality in the rest of the page. You can still drop an image on
the editor, but now it doesn't prompt the user to do so.

Hack taken from https://github.com/summernote/summernote/issues/187