Reverts select box time input
The select box was added in 72f4e7bbc1ce7ec4594fe80d75d19f9dbbcfa3ab to
try and address some usability issues. We were using 15 minute
increments, but this does not work for movie times. We've decided to
revert that change and go back to using the input mask for the time
being.