Event rendering performance improvements
There were some pretty significant rendering performance issues when
trying to display > 100 events at a time. It turned out we were actually
rendering all of them, and only displaying a subset, hiding the
remaining with CSS. This addresses that issue by not rendering
everything, and only doing so if a user chooses to render them.