Sets first day of Pikaday calendar to Sunday
ember-pikaday does not support changing the first day of the calendar,
but Pikaday the library does. This switches over to my own fork of the
addon which allows you to set the firstDay property.

I opened a PR into the main repo so we may be able to switch back to
that at some point. https://github.com/edgycircle/ember-pikaday/pull/41