Fixed a hack with the store and component test
The component test was originally failing for this after upgrading to
the latest version of Ember. I added a hack to ensure the store was
present. After digging into a bit more, I noticed that there was a
deprecation warning related to changing the component state within
didInsertElement. This fixes it by moving that logic to the route where
the store will be available.