Fixes scroll position issue when navigating
This addresses an issue where the scroll position was not being
maintained when navigating from the event index listing to a detailed
event show page. It was a frustrating UX and not something that
typically happens with non-single-page-apps. This addresses it by
maintaining the scroll position on the events index singleton controller
and using that value when the index is loaded.