Allows user to crop existing images
This addresses a UX issue where users were not able to edit an image
once they had left the initial event creation step. By storing the
original image file on the model, we are able to persist it between
steps and use it to show the cropper instead of having the user
re-attach the image.