Getting app ready for 2.0 development
The current production Ember app is mounted within a Rails app at /events,
so the root of the Ember app (/) is actually the Events sub-section.

This also removes the global filters from the events section and makes
it only available when viewing the list of events. This is something
we'll be implementing across the application, and I did it here because
it was broken after moving events around.