Change permalinks to use All Communities
Client wants to be able to share a link like /events/first-friday which
is a custom category grouping that will aggregate events with the "First
friday" category. The problem is that if a user is logged in, events
will only be filtered by their community. If they are logged into a
community that does not have any "First friday" events, they will not
see them. By defaulting to all communities, they will see them now.