Set permalink default dates to "this month"
Client wants to be able to share a link like /events/first-friday which
is a custom category grouping that will aggregate events with the "First
friday" category. The problem is that if a user is logged in, events
will only be filtered by today. If there are no "First friday" events
for today, they will not see them. By defaulting to this month, they
will see them now.