Adds event tracking to Intercom integration
We are already capturing this data in Mixpanel, but are adding it to
Intercom for a couple reasons:

1. We'd like to be able to message users who have used the event filter
and created an event, and ask for their feedback.
2. We'd like to compare Intercom and Mixpanel and see if there's a
reason to use one tool over the other.

Note that the event names are different between Mixpanel and Intercom.
Intercom forces lowercase words separated by dashes. We should probably
use this convention for Mixpanel too, but I didn't want to change the
tracking while it's live.