Adds news filter
This adds the ability to filter news by query, location, and
publication. There's a ton of duplication in the filters now, and I'd
like to spend some time DRYing it up, but I'm holding off on doing that
now until we get the market filter (the last one) in place.