Changes sessions to use ember-simple-auth
The Rails API uses the Devise ruby gem to manage users.
ember-simple-auth has support for Devise, so we're using that almost out
of the box.

In development mode, you can log in with any email address and password.
In staging, you will have to log in with embertest@subtext.org:clearauth
until we get user registration completed.