Adds market contact info
The market contact info is loaded async from the server in a request
that is separate from the original GET /market_posts/:id request. The
reason for this is because we don't want everyone's contact info to be
indexed by Google when we add pre-render support. This is similar to how
Craigslist handles contact info.