Fixes issues with showing old comments on new content
The component was not resetting the comments array when navigating
between pages, this caused it to keep pushing content into the existing
array and showing comments that were for other content. It was also
trying to load comments for new content (i.e. previewing a new Event),
and was making an invalid API request.