Updates event query filter to search with changes
This is similar to the news query input where it updates the search
results when the user is finished typing.

In testing this, I noticed an issue where the contents of the input box
would be overwritten if a user continues typing while an API request is
being made. I'm not totally sure what is causing this but want to come
back to it later if it's a high priority issue.