Clicking publication links filters by that publication
This adds a link to the news card and detail pages that allows a user to
click on a link and view all the news articles for that publication.

The feature request always wanted to be able to search across all
locations when viewing a specific publication. This does not include
that functionality because I was not sure how to do that with the API.
We can add it later.