Adds my postings to the user dashboard
This makes all of a user's postings available on their dashboard. They
can view and edit them (talks cannot be edited) and see some statistics.

This includes the basic layout and functionality, but does not include
pagination or sorting. I'm not sure what we want to do for pagination or
sorting so I'm pushing that off until we can chat about it.