Fixes a duplicate ID ember error on events show page
We were passing an attribute "id" to the edit-content-link component,
but that was setting the ID of the element. This was causing it to
render multiple elements on the page with the same ID, and Ember did not
like this. By changing it to contentId, we get around this issue.