Adds similar/suggested content to all show pages
The similar content is a group of suggested pieces of content that are
related to the page a user is on. The idea is that we want to show
related content so that users will be more engaged with the site.

Similar content is not shown when previewing content because it dpends
on the API receiving its content id, which does not exist at that point.