Fixes custom header for non ember data requests
This fixes an issue where the custom header was only being sent when
making API requests with the ember data store. Manual ajax() requests
were not sending the header because they do not use the store. This
addresses the issue by overriding the devise authorizer, and manually
setting the header there. Now all requests going out contain the custom
header.