Fixes most WYSIWYG links
This fixes an issue where if someone added a link to a URL like
'www.google.com', it was invalid because it didn't contain the protocol.
This naively appends 'http://' to the start of the link to fix it for
those cases.

A better solution would be to validate the URL entirely and make sure
it's valid before saving.