Scroll user to the comment on a talk page
Talks consist of a parent comment and a collection of child comments.
When a user visits a talk show page, we are now taking them to the
parent comment, and scrolling down to the child comment that they
clicked on.

Loading comments was moved to the talk show route because we need to
scroll them down the page as soon as possible, and not later on when the
comment promise loads.