Reset page when going to main index pages
This fixes an issue due to how query params work in Ember. By default
everything is sticky, meaning if you're on page 2 of the events section
and then go to the news section, when you come back to the events
section, you will still be on page 2. This fixes it so you're always
brought to page 1.