Improves UI for news images on card and show page
The images coming in for news articles are all different shapes and
sizes, and the UI did not account for that. It was causing most images
to look terrible because they were cropping people's faces out and
stuff like that. This fixes it to show the whole image, at the expense
of potentially making really small images look a little worse than
before because they are now displayed at their full size.