Fixes content show page back buttons
The back button was hardcoded to always go back to the index page for
the content you were on. Meaning if you were on a news show page, back
always took you to the news index page. This doesn't work because we can
also get to content pages from the main index listing, as well as the
dashboard.

This introduces a service that tracks the route you came in from, and
uses that to link back to the previous page.