Fixes things related to market id change
The market post primary key was recently changed from id to content_id.
This broke things like the related promotion, flagging content, and
similar content. This aliases the content_id to be id, which is really
the content_id. Slightly confusing, but what it now means is that
model.get('id') and model.get('contentId') now both return content_id,
which is what we want.