Fixes market cards on mobile
The heigh:100% style is only needed on desktop because the cards have a
fixed height. On mobile it's not important because the card height
collapses to fit the content. There must be some incompatibility with
Safari on iOS and setting the height on a link, so removing it fixes the
layout.