Fixes updating current user properties
The API is doing a check that the current user's database ID matches the
userId property that we're sending. We were sending a string and the
database ID is a number, so the comparison is failing. The userId should
probably be a number in the Ember app, so this fixes it.