Refactors refresh param functionality
The 'r' query parameter was originally intended to make it so we could
change the query parameters without refreshing the page. This let us
make multiple filter menu selections and then refresh the model only
when a button was clicked. We have since removed the button and no
longer need this behavior. This makes the code a lot cleaner and
easier to understand.