Fixes event scroll position when paginating
The scroll position was being maintained by the view so that when a user
clicks into an event and goes back, they are brought to the same place
on the page. This was breaking pagination because they were brought to
the bottom of the page. This fixes it by making a "global" position
service that can be set anywhere in the app.