Adds email confirmation to user registration
Previously, when a user registered, they were automatically signed into
the application. Now they must confirm their email address before
logging in. When registering, they'll receive an email with a link that
goes to /sign_up/confirm/:token which will verify the token and then log
them in if it is correct.

If the token is invalid or the API returns an error for any reason, the
user is taken to an error page.