Disables changing email address
There is some confusing behavior with changing an email address right
now because the API requires confirmation when changing it. When the API
response comes back after changing an email address, the old address is
sent. This causes the UI to immediately show the old address. We also
don't show a message to the user that tells them they need to confirm
their new email.

Pulling it out for now until we can come up with a better solution which
will require coordination between UI and API.