Changes logic for market extended reach
Previously, extended reach was off by default, and users had to opt into
it. Now it is on by default and users have to turn it off. The logic is
sort of reversed now because the user thinks that they are turning on
"My Town Only", when in reality they are turning off "Extended Reach".