Allows users to remove images when editing markets
Images can now be removed when editing a market post by flagging an
image as "to be deleted". We don't have a way to roll it back in the UI
yet, but if the user refreshes before saving, the images will not be
deleted until the market post is save.