Adds support back for blank end time
In order to support and end date, but no end time, we need to introduce
an additional attribute to the schedule model. The naming isn't great,
but these attributes now represent:

* startsAt - date of the first event instance and start time for all
instances
* endsAt - end time for all event instances
* endDate - end date for the last event instance

I've also changed the default end time to be null instead of "10:00 am"
because it was preventing the app from accurately displaying "no end
time" condition when editing an event.