Fixes issue with editing recurring events
If a recurring event is created and doesn't have the daysOfWeek or
weeksOfMonth arrays populated, the API converts it to nil and it's send
back as null when editing. It turns out that getWithDefault() only works
when the value is undefined, not when it's null. So we were using null
in places where we wanted to have empty arrays.