Fixes an issue with removing schedules while creating
We were sending all schedules to the API, and the ones that were not
supposed to be created had a "_remove" flag set. Since they don't exist
on the API side yet, it was causing the API to raise an exception. This
fixes the issue by preventing schedules from being sent that don't need
to be.