Fixes UX issue with resetting overrides (exclusions)
This fixes a UX issue where overrides (excluded dates) were reset any
time a value was changed on a schedule. We don't need to or want to
reset the overrides when changing things like time, subtitle, or
presenter, so we're not doing this any more. We're now only removing
overrides that are outside of the current date range, and leaving the
others in place.