Removes unneeded "autoload" code from weather component
The autoload=true was just a hack so that the unit tests didn't try and
make an AJAX request to the API during the test. This updates the test
to rely on the mirage server to handle the API request, allowing the
template to be rendered like it is in the browser.