Fixes UX issue with discarding changes
This adds some additional behavior around creating and editing content.
When someone is creating content, if they navigate to another page,
they will be prompted to acknowledge that they want to discard their
content. When they are editing content, if they navigate to another page,
they will be prompted to acknowledge that they want to discard their
changes.

This also adds a "Discard Changes" button at the bottom of the form
when editing.