Fixes issue after saving new avatar in iOS safari
This addresses an issue where the cropUpdated() callback was triggered
after an image was already saved. This only seems to happen in Safari
for iOS, but was causing the user avatar image to go blank after it was
saved. This fixes it by only running the callback if the data it needs
is available.