Improves form error UX
This improves the form error user experience by making it more obvious
when there are errors.

1. Now when a user submits the form, the page scrolls up to the first
error in the form.
2. Error text has been moved next to the label instead of below the
input in cases where this improves the readability.
3. Error messages are removed when a user fixes the issue rather than
when they resubmit the form. For example, if you submit the event form
without a title, the error will be removed as soon as you add the title.