Improves event location UX
This improves the user experience of adding a location when creating or
editing an event. Previously, we showed all the location form fields,
but this was overwhelming users. Now we only show the location name
field, optimizing for existing locations. Users can still add new
locations, but they will do so after the search results have been
returned.