Fixes rolling back market place images
Ember data doesn't automatically rollback relationships, so we need to
manually roll it back. This addresses an issue where a user might delete
a market image and discard the changes, and then decide to edit the post
again.

This also addresses an issue where deleting an image and discarding
changes would not display the image once it was rolled back

Lastly, this addresses an issue where the discard confirmation wouldn't
pop up if the user only changed an image since we weren't detecting
changes to the images.