Fixes rolling back event schedules
Ember data doesn't automatically rollback relationships, so we need
to manually roll it back. This addresses an issue where a user might
delete or edit an event schedule (date) and discard the changes, and
then decide to edit the event again.

This also addresses an issue where the discard confirmation wouldn't pop
up if the user only changed a schedule since we weren't detecting
changes to the schedules.