Refactors how mixpanel tracking arguments are defined
This refactor removes the implicit, positional arguments in favor of
an explicit structured data object. This should make it easier to know
what each component or template is tracking, and make it easier to track
new events.

One bigger change here is that we can now support templates that need to
track different types of events, as opposed to the tracking the same
event with different properties.